+++ src/main/java/niv/flowstone/mixin/LavaFluidMixin.java
@@ -1,7 +1,5 @@
 package niv.flowstone.mixin;
 
-import static niv.flowstone.FlowstoneGenerator.replace;
-
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Redirect;
@@ -10,6 +8,7 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.LevelAccessor;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.material.LavaFluid;
+import niv.flowstone.Flowstone;
 
 @Mixin(LavaFluid.class)
 public class LavaFluidMixin {
@@ -25,6 +24,6 @@ public class LavaFluidMixin {
             at = @At(value = "INVOKE", //
                     target = LEVEL_ACCESSOR + "setBlock(" + BLOCK_POS + BLOCK_STATE + "I" + ")Z"))
     public boolean setBlockStateProxy(LevelAccessor level, BlockPos pos, BlockState state, int flags) {
-        return level.setBlock(pos, replace(level, state), flags);
+        return level.setBlock(pos, Flowstone.replace(level, pos, state), flags);
     }
 }